creating a nicer ABNSearch::ABN class
