+++ Gemfile
@@ -2,3 +2,10 @@ source "http://rubygems.org"
 
 # Specify your gem's dependencies in abnsearch.gemspec
 gemspec
+
+gem 'savon',      '~2.11.1'
+gem 'nokogiri',   '~1.6.6.2'
+
+group :development do
+  gem 'yard',     '~0.8.7.4'
+end
